requirements.txt
altgraph==0.17.2
boto3==1.36.2
botocore==1.36.2
certifi==2024.2.2
charset-normalizer==3.3.2
cramjam==2.9.1
cx_Oracle==8.3.0
dateparser==1.2.0
fastparquet==2024.11.0
fsspec==2024.12.0
future==0.18.2
idna==3.6
jmespath==1.0.1
macholib==1.15.2
numpy==1.26.4
packaging==23.2
pandas==2.2.0
patsy==0.5.6
pyarrow==19.0.0
pycryptodome==3.20.0
pykalman==0.9.7
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
regex==2023.12.25
requests==2.31.0
s3transfer==0.11.1
scipy==1.12.0
six==1.16.0
statsmodels==0.14.1
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.20
websocket-client==1.7.0
